Refactor commands module into submodules
