+++ src/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod particle;
+pub mod foil;
+pub mod state;
+pub mod dispatcher;
+
+pub use dispatcher::process_command;
